Añadido metodo para crear tabla en Derby con el ID, tipo de objeto y nombre.
Restructurados los ResourceManagerWrapper para que hereden de una clase abstracta común con los métodos compartidos. Los métodos se definen protected para poder ser sobreescritos pero no accesible desde fuera (deben ser funciones auxiliares internas).

Los metodos implementados son los relativos a la obtención de las propiedades a procesar.

Se empieza a desarrollar la resolución de colecciones, mapas y arrays. Se determina el generico utilizado en los mismos.

+++ lib/Neo4j/java/org/neo4j/impl/nioneo/store/Store.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2002-2008 "Neo Technology,"
+ * Copyright (c) 2002-2009 "Neo Technology,"
  *     Network Engine for Objects in Lund AB [http://neotechnology.com]
  *
  * This file is part of Neo4j.
@@ -36,7 +36,7 @@ public interface Store
      */
     public int nextId();
     
-    public int getHighestPossibleIdInUse();
+    public long getHighestPossibleIdInUse();
 
-    public int getNumberOfIdsInUse();
+    public long getNumberOfIdsInUse();
 }
