+++ lib/Neo4j/java/org/neo4j/impl/persistence/IdGenerator.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2002-2008 "Neo Technology,"
+ * Copyright (c) 2002-2009 "Neo Technology,"
  *     Network Engine for Objects in Lund AB [http://neotechnology.com]
  *
  * This file is part of Neo4j.
@@ -47,12 +47,12 @@ public class IdGenerator
         return getPersistenceSource().nextId( clazz );
     }
 
-    public int getHighestPossibleIdInUse( Class<?> clazz )
+    public long getHighestPossibleIdInUse( Class<?> clazz )
     {
         return getPersistenceSource().getHighestPossibleIdInUse( clazz );
     }
 
-    public int getNumberOfIdsInUse( Class<?> clazz )
+    public long getNumberOfIdsInUse( Class<?> clazz )
     {
         return getPersistenceSource().getNumberOfIdsInUse( clazz );
     }
